extract functions from: `utils/utils.js`
