--- src/app/g3w-ol/controls/utils.js
@@ -1,8 +1,0 @@
-/**
- * @file
- * @deprecated since v3.8. Will be removed in v3.9. Use `utils/ol.controls` instead.
- */
-
-import utils from 'utils/ol.controls';
-
-module.exports = utils
