+++ src/utils/basemixin.js
@@ -1,0 +1,5 @@
+import { merge } from './merge';
+
+export function basemixin(destination, source) {
+  return merge(destination.prototype, source);
+};
