+++ src/utils/capitalize_first_letter.js
@@ -1,0 +1,3 @@
+export function capitalize_first_letter(string){
+  return `${string[0].toUpperCase()}${string.slice(1)}`;
+};
