+++ src/utils/convertObjectToUrlParams.js
@@ -1,0 +1,3 @@
+export function convertObjectToUrlParams(params = {}) {
+  return $.param(params)
+};
