+++ src/utils/debounce.js
@@ -1,0 +1,13 @@
+
+/**
+ * build debounce function
+ */
+export function debounce(func, delay=500) {
+  let timeout;
+  return function (...args) {
+    clearTimeout(timeout);
+    timeout = setTimeout(()=>{
+      func(...args)
+    }, delay);
+  };
+};
