+++ src/utils/getRandomColor.js
@@ -1,0 +1,3 @@
+export function getRandomColor(){
+  return `#${((1<<24)*Math.random() | 0).toString(16)}`;
+};
