+++ src/utils/hasOwn.js
@@ -1,0 +1,3 @@
+export function hasOwn(object, key) {
+  return Object.prototype.hasOwnProperty.call(object, key);
+};
