+++ src/utils/inherit.js
@@ -1,0 +1,10 @@
+/**
+ * google closure library impememtation
+ */
+export function inherit(childCtor, parentCtor) {
+  function tempCtor() {}
+  tempCtor.prototype = parentCtor.prototype;
+  childCtor.superClass_ = parentCtor.prototype;
+  childCtor.prototype = new tempCtor();
+  childCtor.prototype.constructor = childCtor;
+};
