+++ src/utils/isEmptyObject.js
@@ -1,0 +1,3 @@
+export function isEmptyObject(obj){
+  return JSON.stringify(obj) === '{}';
+};
