+++ src/utils/merge.js
@@ -1,0 +1,8 @@
+import { hasOwn } from './hasOwn';
+
+export function merge(destination, source) {
+  let key;
+  for (key in source) {
+    if (hasOwn(source, key)) destination[key] = source[key];
+  }
+};
