+++ src/utils/mixin.js
@@ -1,0 +1,7 @@
+import { merge } from './merge';
+
+export function mixin(destination,source) {
+  const sourceInstance = new source;
+  merge(destination, sourceInstance);
+  merge(destination.prototype, source.prototype);
+}
