+++ src/utils/reject.js
@@ -1,0 +1,5 @@
+export function reject(value){
+  const d = $.Deferred();
+  d.reject(value);
+  return d.promise();
+};
