+++ src/utils/resolve.js
@@ -1,0 +1,5 @@
+export function resolve(value){
+  const d = $.Deferred();
+  d.resolve(value);
+  return d.promise();
+};
