+++ src/utils/splitContextAndMethod.js
@@ -1,0 +1,7 @@
+export function splitContextAndMethod(string='') {
+  const [context, method] = string.split(':');
+  return {
+    context,
+    method
+  }
+};
