+++ src/utils/toRawType.js
@@ -1,0 +1,4 @@
+export function toRawType(value) {
+  const _toString = Object.prototype.toString;
+  return _toString.call(value).slice(8, -1)
+};
