+++ src/utils/trimValue.js
@@ -1,0 +1,3 @@
+export function trimValue(value) {
+  return value.replace(/ /g,'');
+}
